requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==3.0.3
django-bootstrap4==1.1.1
django-filter==2.2.0
django-mathfilters==1.0.0
gunicorn==20.0.4
idna==2.9
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mpld3==0.3
multitasking==0.0.9
numpy==1.18.1
pandas==1.0.2
psycopg2-binary==2.8.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
yfinance==0.1.54